rename css and html plugin file names
to match the class name they export. This will make all plugins have same fileName pattern.

export plugins from `plugins/index.ts`
