+++ src/FuseBox.ts
@@ -1,4 +1,4 @@
-import { JSONPlugin } from "./plugins/JSONplugin";
+import { JSONPlugin } from "./plugins/JSONPlugin";
 import { PathMaster } from "./PathMaster";
 import { WorkFlowContext } from "./WorkflowContext";
 import { CollectionSource } from "./CollectionSource";
