+++ src/index.ts
@@ -1,15 +1,4 @@
-export { StylusPlugin } from "./plugins/StylusPlugin";
-export { PostCSS } from "./plugins/PostCSSPlugin";
-export { TypeScriptHelpers } from "./plugins/TypeScriptHelpers";
-export { SVGPlugin } from "./plugins/SVGPlugin";
-export { BabelPlugin } from "./plugins/BabelPlugin";
-export { ChainPlugin } from "./plugins/ChainPlugin";
-export { LESSPlugin } from "./plugins/LESSPlugin";
-export { CSSPlugin } from "./plugins/CSSplugin";
-export { HTMLPlugin } from "./plugins/HTMLplugin";
-export { JSONPlugin } from "./plugins/JSONplugin";
-export { BannerPlugin } from "./plugins/BannerPlugin";
-export { SassPlugin } from "./plugins/SassPlugin";
+export * from './plugins'
 
 export { WorkFlowContext } from "./WorkflowContext";
 export { PathMaster } from "./PathMaster";
