+++ src/plugins/JSONPlugin.ts
